requirements.txt
Flask
gunicorn
numpy
pandas
scikit-learn
pyYAML
-e .